Export Host and Screen, move Container to app's side
