+++ LynxExample/src/components/StackContainerWithDynamicRouteConfigs.tsx
