+++ LynxExample/src/components/StackNavigationButtons.tsx
