+++ LynxExample/src/components/TemplateScreen.tsx
