+++ LynxExample/src/contexts/StackNavigationContext.ts
