+++ LynxExample/src/hooks/useParentNavigationEffect.ts
