+++ LynxExample/src/hooks/useStackNavigationContext.ts
