+++ LynxExample/src/hooks/useStackOperationMethods.ts
