+++ LynxExample/src/hooks/useStackRouteConfigContext.ts
