+++ LynxExample/src/utils/id-generator.ts
