+++ LynxExample/src/utils/reducer.ts
