+++ package.json
@@ -1,40 +1,41 @@
 {
-  "name": "LynxScreens",
-  "version": "1.0.0",
+  "name": "lynx-screens",
+  "version": "0.1.0",
+  "description": "Native stack navigation primitives for Lynx, modeled after react-native-screens",
+  "license": "MIT",
   "type": "module",
+  "main": "./src/index.ts",
+  "files": [
+    "android",
+    "generated",
+    "ios",
+    "src",
+    "lynx.lib.json",
+    "README.md"
+  ],
   "scripts": {
-    "build": "rspeedy build",
-    "dev": "rspeedy dev",
+    "codegen": "lynx-autolink-codegen",
     "format": "prettier --write .",
-    "lint": "eslint .",
-    "preview": "rspeedy preview",
-    "test": "vitest run"
+    "lint": "eslint ."
   },
-  "dependencies": {
-    "@lynx-js/react": "^0.112.5"
+  "peerDependencies": {
+    "@lynx-js/react": "*"
   },
   "devDependencies": {
     "@eslint/js": "^9.32.0",
-    "@lynx-js/preact-devtools": "^5.0.1-6664329",
-    "@lynx-js/qrcode-rsbuild-plugin": "^0.4.1",
-    "@lynx-js/react-rsbuild-plugin": "^0.10.13",
-    "@lynx-js/rspeedy": "^0.11.0",
+    "@lynx-js/autolink-codegen": "^0.2.1",
+    "@lynx-js/react": "^0.112.5",
     "@lynx-js/types": "3.4.11",
-    "@rsbuild/plugin-type-check": "1.2.4",
-    "@testing-library/jest-dom": "^6.8.0",
     "@types/react": "^18.3.23",
     "eslint": "^9.32.0",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.20",
     "globals": "^16.3.0",
-    "jsdom": "^26.1.0",
     "prettier": "^3.6.2",
     "typescript": "~5.9.2",
-    "typescript-eslint": "^8.38.0",
-    "vitest": "^3.2.4"
+    "typescript-eslint": "^8.38.0"
   },
   "engines": {
     "node": ">=18"
-  },
-  "private": true
+  }
 }
