+++ src/index.ts
@@ -1,0 +1,6 @@
+import './lynx-elements';
+
+export { StackHostNativeComponent } from './native_components/StackHostNativeComponent';
+export { StackScreenNativeComponent } from './native_components/StackScreenNativeComponent';
+
+export * from './types/StackScreen';
